requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
azure-storage-blob
azure-identity
jinja2
python-dotenv
